feat(ui/element): seprate mixins
